fix: 修复数字解析区域依赖和 tick 缩放溢出问题
UgcParser/SusParser: TryParse 统一使用 CultureInfo.InvariantCulture
UgcGenerator/SusGenerator: ScaleUp 乘法提升为 long 防溢出

+++ generator/chu/SusGenerator.cs
@@ -52,7 +52,7 @@ public class SusGenerator : IGenerator<IChuChart>
 
     private static ChuNote ScaleUp(ChuNote n)
     {
-        int s(int v) => v * SusTpb / (C2sRsl / 4);
+        int s(int v) => (int)((long)v * SusTpb / (C2sRsl / 4));
         return new ChuNote
         {
             Type = n.Type, Measure = n.Measure, Offset = s(n.Offset),
